requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-docx==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Bcrypt==1.0.1
requests==2.33.1
pypdf==6.10.2